Add types and refactor imports to use new types file
